Move and rename namespace and assembly definition
