+++ Assets/Source/DataDrivenEventTrigger/IEventTriggerDataForTData.cs
@@ -1,4 +1,4 @@
-namespace A3.Event
+namespace A3.DataDrivenEventTrigger
 {
     public interface IEventTriggerData<TData>
     {
