+++ Assets/Source/DataDrivenEventTrigger/IEventTriggerDataForTData.cs.meta
